Atomic PCI-space accesses
These bugs would be very difficult to debug.  If the lock turns out to
be too expensive for bulk PCI ops, we can expose the lock and make some
__pci* versions.

+++ kern/arch/x86/pci.h
@@ -375,12 +375,12 @@ uint32_t pci_config_addr(uint8_t bus, uint8_t dev, uint8_t func, uint32_t reg);
 uint32_t pci_read32(uint8_t bus, uint8_t dev, uint8_t func, uint32_t offset);
 void pci_write32(uint8_t bus, uint8_t dev, uint8_t func, uint32_t offset,
                  uint32_t value);
-uint32_t pci_read16(uint8_t bus, uint8_t dev, uint8_t func, uint32_t offset);
+uint16_t pci_read16(uint8_t bus, uint8_t dev, uint8_t func, uint32_t offset);
 void pci_write16(uint8_t bus, uint8_t dev, uint8_t func, uint32_t offset,
-                 uint32_t value);
-uint32_t pci_read8(uint8_t bus, uint8_t dev, uint8_t func, uint32_t offset);
+                 uint16_t value);
+uint8_t pci_read8(uint8_t bus, uint8_t dev, uint8_t func, uint32_t offset);
 void pci_write8(uint8_t bus, uint8_t dev, uint8_t func, uint32_t offset,
-                uint32_t value);
+                uint8_t value);
 uint32_t pcidev_read32(struct pci_device *pcidev, uint32_t offset);
 void pcidev_write32(struct pci_device *pcidev, uint32_t offset, uint32_t value);
 uint16_t pcidev_read16(struct pci_device *pcidev, uint32_t offset);
